+++ .eslintrc.js
@@ -18,5 +18,6 @@ module.exports = {
     'react',
   ],
   rules: {
+    'linebreak-style': ['error', 'windows'],
   },
 };
