+++ src/components/BG/bg.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import './bg.png';
+
+function BG() {
+  return (
+    <div className="page-background" />
+  );
+}
+
+export default BG;
