+++ src/components/BG/index.jsx
@@ -1,10 +1,3 @@
-import React from 'react';
-import './bg.png';
-
-function BG() {
-  return (
-    <div className="page-background" />
-  );
-}
+import BG from './bg';
 
 export default BG;
