+++ src/components/header/index.jsx
@@ -1,22 +1,3 @@
-import React from 'react';
-import Logo from '../logo';
-import LagnSwithcher from '../lagnSwithcher';
+import Header from './header';
 
-function header() {
-  return (
-    <header className="header">
-      <div className="container">
-        <div className="header__inner">
-          <div className="header__left">
-            <Logo />
-          </div>
-          <div className="header__right">
-            <LagnSwithcher />
-          </div>
-        </div>
-      </div>
-    </header>
-  );
-}
-
-export default header;
+export default Header;
