+++ src/components/main/main.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function main() {
+  return (
+    <main className="main">
+      <div className="container">
+        <div className="main__inner" />
+      </div>
+    </main>
+  );
+}
+
+export default main;
