+++ src/index.jsx
@@ -19,9 +19,7 @@ function App() {
 }
 
 ReactDOM.render(
-  // <React.StrictMode>
-  <App />, // </React.StrictMode>,
-
+  <App />,
   document.getElementById('root'),
 );
